+++ post/index.js
@@ -3,7 +3,7 @@ const postChainTransaction = require('./post_chain_transaction');
 const postClosedMessage = require('./post_closed_message');
 const postOpenMessage = require('./post_open_message');
 const postSettledInvoice = require('./post_settled_invoice');
-const postUpdatedBackups = require('./post_updated_backups');
+const postUpdatedBackup = require('./post_updated_backup');
 const sendFile = require('./send_file');
 const sendMessage = require('./send_message');
 
@@ -13,7 +13,7 @@ module.exports = {
   postClosedMessage,
   postOpenMessage,
   postSettledInvoice,
-  postUpdatedBackups,
+  postUpdatedBackup,
   sendFile,
   sendMessage,
 };