+++ test/post/test_post_updated_backup.js
@@ -1,0 +1,77 @@
+const {test} = require('tap');
+
+const {postUpdatedBackup} = require('./../../post');
+
+const makeArgs = overrides => {
+  const args = {
+    backup: '00',
+    id: 1,
+    key: 'key',
+    node: {
+      alias: 'alias',
+      public_key: Buffer.alloc(33).toString('hex'),
+    },
+    request: ({}, cbk) => {
+      return cbk(null, {statusCode: 200});
+    },
+  };
+
+  Object.keys(overrides).forEach(k => args[k] = overrides[k]);
+
+  return args;
+};
+
+const tests = [
+  {
+    args: makeArgs({backup: undefined}),
+    description: 'Posting an updated backup requires a backup',
+    error: [400, 'ExpectedBackupFileToPostUpdatedBackup'],
+  },
+  {
+    args: makeArgs({id: undefined}),
+    description: 'Posting an updated backup requires a user id',
+    error: [400, 'ExpectedIdToPostUpdatedBackup'],
+  },
+  {
+    args: makeArgs({key: undefined}),
+    description: 'Posting an updated backup requires an api key',
+    error: [400, 'ExpectedApiKeyToPostUpdatedBackup'],
+  },
+  {
+    args: makeArgs({node: undefined}),
+    description: 'Posting an updated backup requires node details',
+    error: [400, 'ExpectedNodeToPostUpdatedBackup'],
+  },
+  {
+    args: makeArgs({request: undefined}),
+    description: 'Posting an updated backup requires a request function',
+    error: [400, 'ExpectedRequestFunctionToPostUpdatedBackup'],
+  },
+  {
+    args: {
+      backup: '00',
+      id: 1,
+      key: 'key',
+      node: {
+        alias: 'alias',
+        public_key: Buffer.alloc(33).toString('hex'),
+      },
+      request: ({}, cbk) => {
+        return cbk(null, {statusCode: 200});
+      },
+    },
+    description: 'An updated backup is posted',
+  },
+];
+
+tests.forEach(({args, description, error, expected}) => {
+  return test(description, async ({end, equal, rejects}) => {
+    if (!!error) {
+      await rejects(postUpdatedBackup(args), error, 'Got expected error');
+    } else {
+      await postUpdatedBackup(args);
+    }
+
+    return end();
+  });
+});